Use slices instead of VAddr
This changes `paddr_to_kernel_vaddr` to `paddr_to_slice` which
additionally takes a length. This more accurately expresses what we
depend on. Also, paddr_to_slice returns an option, allowing users to
reject translations altogether.

This changes the Module object to provide PAddr instead of VAddr,
allowing us to get rid of VAddr altogether. I feel this is more
appropriate because this library shoudn't be concerned with the virtual
addresses of the modules, just the virtual addresses of the information
provided within the multiboot struct. Alternatively we could provide a
slice of bytes(&[u8]).

The big outstanding question to me is what to do about cstrings. I think
it probably is appropriate to ask for the translation for each byte -
even though this is inefficient it is safer. Efficiency isn't exactly a
prime concern when dealing with the multiboot information anyhow.

Signed-off-by: Dan Schatzberg <schatzberg.dan@gmail.com>